utils: add Opaque object that implements casting methods
