+++ src/platform_impl/web/web_sys/fullscreen.rs
@@ -11,6 +11,7 @@ use super::super::main_thread::MainThreadMarker;
 use super::super::monitor::{self, ScreenDetailed};
 use crate::monitor::Fullscreen;
 use crate::platform_impl::MonitorHandle;
+use crate::utils::OpaqueObject;
 
 pub(crate) fn request_fullscreen(
     main_thread: MainThreadMarker,
